Перенёс вспомогательные функции в отдельный файл
